move constants to its own file
