+++ modules/client/src/index.ts
@@ -5,3 +5,4 @@ export * from "./client-common";
 export * from "./multisig";
 export * from "./interfaces";
 export * from "./types";
+export * from "./constants";
